Disable support for migrating existing configuration
These blocks don't survive container restarts. Need to be refactored later.

+++ entrypoint.sh
@@ -12,23 +12,15 @@ if [[ "$RESET_SEED" == "true" ]]; then
 fi
 
 SETTINGS_FILE="$APP_DIR/MoriaServerConfig.ini"
-if [[ -f $SETTINGS_FILE ]]; then
-    echo "Existing configuration file found, moving to $CONFIG_DIR"
-    mv $SETTINGS_FILE "$CONFIG_DIR/MoriaServerConfig.ini"
-fi
+envsubst < /MoriaServerConfig.ini.tmpl > "$CONFIG_DIR/MoriaServerConfig.ini"
+ln -sf "$CONFIG_DIR/MoriaServerConfig.ini" $SETTINGS_FILE
 
-SAVE_DIR="$APP_DIR/Moria/Saved"
-if [[ -d $SAVE_DIR ]]; then
+SAVE_DIR="$APP_DIR/Moria"
+if [[ -d $SAVE_DIR && ! -L $SAVE_DIR ]]; then
     echo "Moving existing save to $DATA_DIR"
     mv "$APP_DIR/Moria" "$DATA_DIR"
-else
-    echo "No existing save file found"
-    mkdir -p "$APP_DIR/Moria"
 fi
-
-envsubst < /MoriaServerConfig.ini.tmpl > "$CONFIG_DIR/MoriaServerConfig.ini"
-ln -s "$DATA_DIR" $SAVE_DIR
-ln -s "$CONFIG_DIR/MoriaServerConfig.ini" $SETTINGS_FILE
+ln -sf "$DATA_DIR" $SAVE_DIR
 
 echo "Starting fake screen"
 rm -f /tmp/.X0-lock 2>&1
